requirements.txt
telethon>=1.35.0
sqlalchemy>=2.0.49
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=23.0.0
alembic>=1.13.0
